Modified Scripts and Made Corrections
