+++ 2D-Snake-Game/Assets/Scripts/Camera/CameraShake.cs
@@ -2,51 +2,45 @@ using System.Collections;
 using UnityEngine;
 using Cinemachine;
 
-public class CameraShake : MonoBehaviour
-{
-    [Header("Cinemachine Virtual Camera")]
-    [SerializeField] private CinemachineVirtualCamera virtualCamera;
-    [Header("Camera Shake Amplitude")]
-    [SerializeField] private float amplitude = 15f;
-    [Header("Camera Shake Time")]
-    [SerializeField] private float shakeTime = 0.2f;
-
-    private CinemachineBasicMultiChannelPerlin noise;
-
-    public static CameraShake Instance { get { return instance; } }
-    private static CameraShake instance;
-     
-    private Coroutine shake; // to store coroutine
-    private WaitForSeconds waitTime;
-    private void Awake()
+    public class CameraShake : MonoBehaviour
     {
-        if (instance != null) 
+        [SerializeField] private CinemachineVirtualCamera virtualCamera;
+        [SerializeField] private float amplitude = 15f;
+        [SerializeField] private float shakeTime = 0.2f;
+
+        private CinemachineBasicMultiChannelPerlin noise;
+
+        public static CameraShake Instance { get { return instance; } }
+        private static CameraShake instance;
+
+        private Coroutine shakeCoroutine;
+        private WaitForSeconds waitTime;
+        private void Awake()
         {
-            Destroy(gameObject);
+            if (instance != null)
+            {
+                Destroy(gameObject);
+            }
+            instance = this;
         }
-        instance = this;
-    }
-    // Start is called before the first frame update
-    void Start()
-    {
-        noise = virtualCamera.GetCinemachineComponent<CinemachineBasicMultiChannelPerlin>();
-        waitTime = new(shakeTime);
-       
+        void Start()
+        {
+            noise = virtualCamera.GetCinemachineComponent<CinemachineBasicMultiChannelPerlin>();
+            waitTime = new(shakeTime);
+
+        }
+        public void ShakeCamera()
+        {
+            if (shakeCoroutine != null)
+                StopCoroutine(shakeCoroutine);
+
+            shakeCoroutine = StartCoroutine(ShakeCoroutine());
+        }
+        IEnumerator ShakeCoroutine()
+        {
+            noise.m_AmplitudeGain = amplitude;
+            yield return waitTime;
+            noise.m_AmplitudeGain = 0;
+        }
     }
 
-    public void ShakeCamera()
-    {
-        // check for running coroutine
-        if(shake != null)
-            StopCoroutine(shake); // if running, then stop
-        // start next coroutine
-        shake = StartCoroutine(ShakeCoroutine()); 
-    }
-
-    IEnumerator ShakeCoroutine()
-    {
-        noise.m_AmplitudeGain = amplitude;
-        yield return waitTime;
-        noise.m_AmplitudeGain = 0;
-    }
-}
