+++ 2D-Snake-Game/Assets/Scripts/Consumables/MassBurner.cs.meta
