+++ 2D-Snake-Game/Assets/Scripts/Consumables/MassGainer.cs
@@ -1,0 +1,16 @@
+public class MassGainer : Food
+{
+        protected override void ConsumeEffect(SnakeController snake)
+        {
+            for (int i = 0; i < amount; i++)
+            {
+                snake.Grow();
+                SoundManager.Instance.Play(Sounds.pickup);
+            }
+            if (snake.scoreController != null)
+            {
+                int score = Score;
+                snake.scoreController.IncreaseScore(score);
+            }
+        }
+}
