--- 2D-Snake-Game/Assets/Scripts/Food/MassGainer.cs
@@ -1,17 +1,0 @@
-public class MassGainer : Food
-{
-    protected override void ConsumeEffect(SnakeController snake)
-    {
-        for (int i = 0; i < amount; i++)
-        {
-            snake.Grow();
-            SoundManager.Instance.Play(Sounds.pickup);
-        }
-        // Call IncreaseScore on the snake's ScoreController
-        if (snake.scoreController != null)
-        {
-            int score = GetScore();
-            snake.scoreController.IncreaseScore(score);
-        }
-    }
-}
