+++ 2D-Snake-Game/Assets/Scripts/Lobby/LobbyController.cs
@@ -4,13 +4,9 @@ using UnityEngine.SceneManagement;
 
 public class LobbyController : MonoBehaviour
 {
-    [Header("Play Button")]
     [SerializeField] private Button buttonPlay;
-    [Header("Quit Button")]
     [SerializeField] private Button buttonQuit;
-    [Header("SingleMode Button")]
     [SerializeField] private Button buttonSingle;
-    [Header("CoOpMode Button")]
     [SerializeField] private Button buttonCoOp;
 
     public GameObject ModeSelection;
@@ -19,33 +15,27 @@ public class LobbyController : MonoBehaviour
     {
         buttonPlay.onClick.AddListener(PlayGame);
         buttonQuit.onClick.AddListener(QuitGame);
-        buttonSingle.onClick.AddListener(SingleMode); 
+        buttonSingle.onClick.AddListener(SingleMode);
         buttonCoOp.onClick.AddListener(CoOp);
     }
-
     private void PlayGame()
     {
         SoundManager.Instance.Play(Sounds.buttonClick);
         ModeSelection.SetActive(true);
     }
-
     private void QuitGame()
     {
         Debug.Log(" Quit Application");
         SoundManager.Instance.Play(Sounds.buttonClick);
         Application.Quit();
     }
-
     private void SingleMode()
     {
-        Time.timeScale = 0.7f;
         SoundManager.Instance.Play(Sounds.buttonClick);
         SceneManager.LoadScene(1);
     }
-
     private void CoOp()
     {
-        Time.timeScale = 0.7f;
         SoundManager.Instance.Play(Sounds.buttonClick);
         SceneManager.LoadScene(2);
     }
