+++ 2D-Snake-Game/Assets/Scripts/Power-Ups/PowerUp.cs
@@ -2,32 +2,28 @@ using UnityEngine;
 
 public enum Power
 {
-    Speed,
-    Shield,
-    Score
+        Speed,
+        Shield,
+        Score
 }
 
 public class PowerUp : Consumable
 {
-    [Header("Type Of Power-Up")]
-    [SerializeField] private Power power;
-    [Header("Power-Up Active Time")]
-    [SerializeField] private float activeTime;
+        [SerializeField] private Power power;
+        [SerializeField] private float activeTime;
 
-    private void OnEnable()
-    {
-        SetPosition();
-    }
-
-    public override void Consume(SnakeController snake)
-    {
-        snake.ActivatePowerUp(power, activeTime);
-        gameObject.SetActive(false);
-    }
-
-    protected override void ResetOnScreenTimer()
-    {
-        base.ResetOnScreenTimer();
-        gameObject.SetActive(false);
-    }
+        private void OnEnable()
+        {
+            SetPosition();
+        }
+        public override void Consume(SnakeController snake)
+        {
+            snake.ActivatePowerUp(power, activeTime);
+            gameObject.SetActive(false);
+        }
+        protected override void ResetOnScreenTimer()
+        {
+            base.ResetOnScreenTimer();
+            gameObject.SetActive(false);
+        }
 }
